feat: expose extrinsics operations as a library (#1181)
* feat: expose extrinsics operations as a library

This commit introduces a library that exposes extrinsics operations as a reusable component.
This makes it easier to integrate and reuse these operations in various projects.


* fix: move  to the extrinsics crate

This commit refines the code to avoid exposing the entire  module


* fix(extrinsics): update docs to initialize BalanceVariant from string

This commit exposes DenominatedBalance and UnitPrefix.

This commit also refactors 3 docs examples to initialize BalanceVariant from string because its fields are private


* fix(extrinsics): update cargo.toml file and run cargo build


* fix(extrinsics): avoid exposing DonomindatedBalance and UnitPrefix


---------
