+++ crates/cargo-contract/Cargo.toml
@@ -19,31 +19,20 @@ include = [
 
 [dependencies]
 contract-build = { version = "3.0.1", path = "../build" }
-contract-metadata = { version = "3.0.1", path = "../metadata" }
+contract-extrinsics = { version = "3.0.1", path = "../extrinsics" }
 contract-transcode = { version = "3.0.1", path = "../transcode" }
 
 anyhow = "1.0.71"
 clap = { version = "4.3.10", features = ["derive", "env"] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-scale = { package = "parity-scale-codec", version = "3.0.0", features = ["derive"] }
 which = "4.4.0"
 colored = "2.0.1"
-serde = { version = "1.0.165", default-features = false, features = ["derive"] }
 serde_json = "1.0.99"
 url = { version = "2.4.0", features = ["serde"] }
-rust_decimal = "1.30"
-
-# dependencies for extrinsics (deploying and calling a contract)
 async-std = { version = "1.12.0", features = ["attributes", "tokio1"] }
-sp-core = "21.0.0"
-sp-runtime = "24.0.0"
-sp-weights = "20.0.0"
-pallet-contracts-primitives = "24.0.0"
-scale-info = "2.8.0"
 subxt = "0.29.0"
 hex = "0.4.3"
-jsonrpsee = { version = "0.18.2", features = ["ws-client"] }
 
 [build-dependencies]
 anyhow = "1.0.71"
