+++ crates/cargo-contract/src/cmd/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2022 Parity Technologies (UK) Ltd.
+// Copyright 2018-2023 Parity Technologies (UK) Ltd.
 // This file is part of cargo-contract.
 //
 // cargo-contract is free software: you can redistribute it and/or modify
@@ -18,7 +18,6 @@ pub mod build;
 pub mod decode;
 pub mod encode;
 pub mod info;
-pub mod runtime_api;
 
 pub(crate) use self::{
     build::{
@@ -28,9 +27,8 @@ pub(crate) use self::{
     decode::DecodeCommand,
     info::InfoCommand,
 };
-mod extrinsics;
 
-pub(crate) use self::extrinsics::{
+pub(crate) use contract_extrinsics::{
     CallCommand,
     ErrorVariant,
     InstantiateCommand,
@@ -38,12 +36,4 @@ pub(crate) use self::extrinsics::{
     UploadCommand,
 };
 
-use subxt::{
-    Config,
-    OnlineClient,
-};
-
 pub use subxt::PolkadotConfig as DefaultConfig;
-type Client = OnlineClient<DefaultConfig>;
-type Balance = u128;
-type CodeHash = <DefaultConfig as Config>::Hash;
