+++ crates/extrinsics/src/call.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Parity Technologies (UK) Ltd.
+// Copyright 2018-2023 Parity Technologies (UK) Ltd.
 // This file is part of cargo-contract.
 //
 // cargo-contract is free software: you can redistribute it and/or modify
@@ -16,33 +16,27 @@
 
 use super::{
     display_contract_exec_result,
+    display_contract_exec_result_debug,
+    display_dry_run_result_warning,
+    events::DisplayEvents,
     prompt_confirm_tx,
+    runtime_api::api,
     state_call,
     submit_extrinsic,
+    Balance,
     BalanceVariant,
     Client,
     ContractMessageTranscoder,
     DefaultConfig,
+    ErrorVariant,
     ExtrinsicOpts,
     PairSigner,
     StorageDeposit,
     TokenMetadata,
+    DEFAULT_KEY_COL_WIDTH,
     MAX_KEY_COL_WIDTH,
 };
 
-use crate::{
-    cmd::{
-        extrinsics::{
-            display_contract_exec_result_debug,
-            display_dry_run_result_warning,
-            events::DisplayEvents,
-            ErrorVariant,
-        },
-        runtime_api::api,
-        Balance,
-    },
-    DEFAULT_KEY_COL_WIDTH,
-};
 use contract_build::name_value_println;
 
 use anyhow::{
