+++ crates/extrinsics/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Parity Technologies (UK) Ltd.
+// Copyright 2018-2023 Parity Technologies (UK) Ltd.
 // This file is part of cargo-contract.
 //
 // cargo-contract is free software: you can redistribute it and/or modify
@@ -20,11 +20,14 @@ mod error;
 mod events;
 mod instantiate;
 mod remove;
+mod runtime_api;
 mod upload;
 
 #[cfg(test)]
 #[cfg(feature = "integration-tests")]
 mod integration_tests;
+
+use subxt::utils::AccountId32;
 
 use anyhow::{
     anyhow,
@@ -46,19 +49,13 @@ use std::{
     path::PathBuf,
 };
 
-use crate::{
-    cmd::{
-        Balance,
-        Client,
-    },
-    DEFAULT_KEY_COL_WIDTH,
-};
-
+use crate::runtime_api::api::{self,};
 use contract_build::{
     name_value_println,
     CrateMetadata,
     Verbosity,
     VerbosityFlags,
+    DEFAULT_KEY_COL_WIDTH,
 };
 use pallet_contracts_primitives::ContractResult;
 use scale::{
@@ -98,6 +95,9 @@ pub use subxt::PolkadotConfig as DefaultConfig;
 pub use upload::UploadCommand;
 
 type PairSigner = tx::PairSigner<DefaultConfig, sr25519::Pair>;
+pub type Client = OnlineClient<DefaultConfig>;
+pub type Balance = u128;
+pub type CodeHash = <DefaultConfig as Config>::Hash;
 
 /// Arguments required for creating and sending an extrinsic to a substrate node.
 #[derive(Clone, Debug, clap::Args)]
@@ -317,7 +317,7 @@ impl WasmCode {
     }
 }
 
-/// Create a new [`PairSigner`] from the given [`sr25519::Pair`].
+/// Create a new `PairSigner` from the given [`sr25519::Pair`].
 pub fn pair_signer(pair: sr25519::Pair) -> PairSigner {
     PairSigner::new(pair)
 }
@@ -467,6 +467,69 @@ pub fn parse_code_hash(input: &str) -> Result<<DefaultConfig as Config>::Hash> {
     let mut arr = [0u8; 32];
     arr.copy_from_slice(&bytes);
     Ok(arr.into())
+}
+
+/// Fetch the contract info from the storage using the provided client.
+pub async fn fetch_contract_info(
+    contract: &AccountId32,
+    client: &Client,
+) -> Result<Option<ContractInfo>> {
+    let info_contract_call = api::storage().contracts().contract_info_of(contract);
+
+    let contract_info_of = client
+        .storage()
+        .at_latest()
+        .await?
+        .fetch(&info_contract_call)
+        .await?;
+
+    match contract_info_of {
+        Some(info_result) => {
+            let convert_trie_id = hex::encode(info_result.trie_id.0);
+            Ok(Some(ContractInfo {
+                trie_id: convert_trie_id,
+                code_hash: info_result.code_hash,
+                storage_items: info_result.storage_items,
+                storage_item_deposit: info_result.storage_item_deposit,
+            }))
+        }
+        None => Ok(None),
+    }
+}
+
+#[derive(serde::Serialize)]
+pub struct ContractInfo {
+    trie_id: String,
+    code_hash: CodeHash,
+    storage_items: u32,
+    storage_item_deposit: Balance,
+}
+
+impl ContractInfo {
+    /// Convert and return contract info in JSON format.
+    pub fn to_json(&self) -> Result<String> {
+        Ok(serde_json::to_string_pretty(self)?)
+    }
+
+    /// Display contract information in a formatted way
+    pub fn basic_display_format_contract_info(&self) {
+        name_value_println!("TrieId", format!("{}", self.trie_id), MAX_KEY_COL_WIDTH);
+        name_value_println!(
+            "Code Hash",
+            format!("{:?}", self.code_hash),
+            MAX_KEY_COL_WIDTH
+        );
+        name_value_println!(
+            "Storage Items",
+            format!("{:?}", self.storage_items),
+            MAX_KEY_COL_WIDTH
+        );
+        name_value_println!(
+            "Storage Deposit",
+            format!("{:?}", self.storage_item_deposit),
+            MAX_KEY_COL_WIDTH
+        );
+    }
 }
 
 /// Copy of `pallet_contracts_primitives::StorageDeposit` which implements `Serialize`,
