+++ crates/transcode/src/lib.rs
@@ -417,7 +417,7 @@ impl ContractMessageTranscoder {
                 data.len(),
                 arg_list_string,
                 encoded_bytes
-            ))
+            ));
         }
         Ok(())
     }
